casting

+++ src/utils.py
@@ -95,10 +95,10 @@ class collateFunction():
                 pad = torch.zeros(self.min_L - L, H, W)
                 lr_batch.append(torch.cat([lrs, pad], dim=0))
                 weights = weights.type(torch.FloatTensor)
-                # print(weights)
-                # print(torch.zeros(self.min_L - L))
-                weights_batch.append(torch.cat([weights, torch.zeros(self.min_L - L)], dim=0))
-                alpha_batch.append(torch.cat([torch.ones(L), torch.zeros(self.min_L - L)], dim=0))
+                weights_batch.append(torch.cat(
+                    [weights, torch.zeros(self.min_L - L, dtype=torch.float64)], dim=0))
+                alpha_batch.append(
+                    torch.cat([torch.ones(L), torch.zeros(self.min_L - L)], dim=0))
 
             hr = imageset['hr']
             if self.train_batch and hr is not None:
